+++ ui/src/app/components/forms.tsx
@@ -1,1 +1,11 @@
 import * as React from 'react';
+
+export const styles = {
+  formSection: {
+    marginTop: '1rem',
+    minWidth: '29rem'
+  }
+};
+
+export const FormSection = ({style = {}, ...props}) =>
+  <div {...props} style={{...styles.formSection, ...style}} />;
